+++ src/entmoot/routes/__init__.py
@@ -1,4 +1,5 @@
+from .admin import AdminController
 from .attributes import AttributeController
 from .entities import EntityController
 
-__all__ = ["EntityController", "AttributeController"]
+__all__ = ["AdminController", "EntityController", "AttributeController"]